+++ patch_subset/cbor/int_utils.cc
@@ -16,13 +16,13 @@ uint32_t IntUtils::ZigZagEncode(int32_t signed_int) {
 }
 
 int32_t IntUtils::ZigZagDecode(uint32_t unsigned_int) {
-  if (unsigned_int & 1) {
+  if (unsigned_int & 1U) {
     if (unsigned_int == UINT32_MAX) {
       return INT32_MIN;
     }
-    return -((unsigned_int + 1) / 2);
+    return (int32_t) -((unsigned_int + 1U) / 2U);
   } else {
-    return unsigned_int / 2;
+    return (int32_t) unsigned_int / 2U;
   }
 }
